Test methods with test cases.

+++ Probe.m
@@ -2,8 +2,8 @@ function [ value ] = Probe(func,variable_number, X, S, probing_length, maxi)
     valPos = X + S.*probing_length;
     valNeg = X - S.*probing_length;
     syms x y 
-    valuePos = subs(func, [x,y], valPos);
-    valueNeg = subs(func, [x,y], valNeg);
+    valuePos = double(subs(func, [x,y], valPos));
+    valueNeg = double(subs(func, [x,y], valNeg));
     if(maxi == 1)
         if(valuePos >= valueNeg)
             value = S;
